Moved class into separate file and instantiate the plugin logic from itself
